Restore
Apologies for the mess. I was requested to restore a particular old version of the book. I backed up the files and worked to restore the version. Once I was done, to be sure everything was as it should be, I deleted all the files. Once pushed, I replaced the original backed up files, so the book should be as it was before the start of this.